Add image pick intent and save photo to Firebase storage
Clicking the image icon at the start of the EditText field opens up the picker intent that lets a user pick a photo and send it. The send button triggers the UploadTask which uploads the selected photo to Firebase storage. Also at the same time it adds a message object to the Realtime Database which calls onChildAdded method inside the ChildEventListener and the realtime update takes place.